DamsoleAIChatbot/requirements.txt
Flask
python-dotenv
openai
requests
